refactoring della soluzione
