+++ createTables.js
@@ -1,0 +1,2 @@
+const executeQuery = require("./mysql.js")
+executeQuery("./createTables.sql")
