--- deleteDb.sql
@@ -1,1 +1,0 @@
-DROP DATABASE ScienzeInGioco;
