+++ package.json
@@ -2,7 +2,6 @@
   "name": "website-db-scripts",
   "version": "1.0.0",
   "description": "",
-  "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
