+++ readme.md
@@ -9,3 +9,13 @@ Prima di eseguire gli script è necessario creare un file di nome `db-config-<en
   "database": ""
 }
 ```
+
+nel caso di sviluppo in laboratorio il file di configurazione deve essere:
+```
+{
+  "host": "192.168.1.231",
+  "user": "scienze_gioco",
+  "password": "admin",
+  "database": "scienze_gioco"
+}
+```
